Change label 'Rules' => 'Models' and separate ModelSelect into two components: ModelSelect and SetGroupSelect
