+++ src/components/2-molecules/SetGroupSelectStyle.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021, RTE (http://www.rte-france.com)
+ * Copyright (c) 2023, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -12,4 +12,13 @@ export const useStyles = makeStyles((theme) => ({
         textAlign: 'right',
     },
     box: {},
+    button: ({ errorInParams }) => ({
+        justifyContent: 'center',
+        display: 'flex',
+        '& .MuiIconButton-root .MuiIconButton-label .MuiSvgIcon-root': {
+            fontSize: '2em',
+            pointerEvents: 'auto',
+            color: errorInParams ? 'red' : 'unset',
+        },
+    }),
 }));
